pyproject.toml
[project]
name = "giraf-core"
version = "0.1.0"
description = "GIRAF Core API — shared domain service for users, organizations, citizens, grades, and pictograms."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2,<6.0",
    "django-ninja>=1.3,<2.0",
    "django-ninja-jwt>=5.3,<6.0",
    "psycopg[binary]>=3.2,<4.0",
    "gunicorn>=23.0,<24.0",
    "django-cors-headers>=4.6,<5.0",
    "django-ninja-extra>=0.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-django>=4.9",
    "factory-boy>=3.3",
    "ruff>=0.9",
    "mypy>=1.14",
    "django-stubs>=5.1",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py312"
line-length = 120
exclude = ["**/migrations/**"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
plugins = ["mypy_django_plugin.main"]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true

[tool.django-stubs]
django_settings_module = "config.settings.test"
